pyproject.toml
[project]
name = "feishu-super"
version = "0.1.0"
description = "Feishu Bitable OpenAPI skill for LLM agents — CRUD with permission guard + rich query DSL"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "typer>=0.12",
    "rich>=13.7",
]

[project.scripts]
feishu-super = "feishu_super.cli:app"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/feishu_super"]
